Remove API-specific Graphics class headers and code duplication. Closes #1422.
